requirements_gpu.txt
numpy
scipy
matplotlib
pandas
Pillow
pytest
tensorflow-gpu>=1.6
flask
